requirements_native.txt
# THEBOT - Dépendances Application Native
# Requirements spécifiques pour l'interface desktop

# Interface native principale
PyQt6>=6.5.0
PyQt6-tools>=6.5.0

# Charts et visualisations natives
PyQt6-Charts>=6.5.0

# Support système
psutil>=5.9.0

# Fallback interface
matplotlib>=3.7.0

# Threading et async pour temps réel
asyncio>=3.4.3

# Icônes et ressources
pillow>=10.0.0

# Notifications système
plyer>=2.1.0

# Configuration avancée
configparser>=5.3.0

# Logging avancé pour debug
loguru>=0.7.0

# Tests spécifiques GUI
pytest-qt>=4.2.0